backend/requirements.txt
asgiref==3.3.1
Django==3.1.4
gunicorn==20.0.4
Pillow==10.2.0
pytz==2020.4
sqlparse==0.4.1

psycopg2==2.8.6
django-filter==2.3.0
djangorestframework==3.11.0

djoser==2.1.0